requirements.txt
discord.py
colorama
